refact handlers
